requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1

# Vision
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
Pillow>=10.3.0
opencv-python>=4.9.0
ultralytics>=8.2.0          # YOLOv8

# FashionCLIP / CLIP
open-clip-torch>=2.24.0

# Vector search
faiss-cpu>=1.8.0            # swap for faiss-gpu on CUDA

# NCF re-ranker
numpy>=1.26.0
scikit-learn>=1.4.0
scipy>=1.13.0

# Data
datasets>=2.19.0            # HuggingFace datasets (Marqo/fashion200k)
huggingface-hub>=0.22.0
requests>=2.31.0
tqdm>=4.66.0
pandas>=2.2.0

# Evaluation
rouge-score>=0.1.2

# Dev
python-dotenv>=1.0.0
pytest>=8.2.0
